Organ Changes and Addition of Our Project.
Okay so I've kinda attached myself to this idea and want to take control
of it and be a total deva so stop me but its honestly too late already
so ye.

Idea 1: I've added Karthik and I's project, and I've changed some of the
file names. Fairly basic, but also kinda controlling. u decide what
happens. I'll try to update this copy of my project as our repo of it
changes also, because idk how to use submodules.

Idea 2: This is just a thought, but wouldn't it be better to organize
classes by class year? I feel like it would be easier to tell who did
what, and that that is more important than the exact time they did it.
Someone looking it rn would say "wow so in 2k15 they were doing this"
whereas someone looking at them organized by graduation year would be
able to easily determine that "my great grandseniors did this"--you get
the idea. I'm just obsessing sorry.